Refactor by packages
