+++ README.md
@@ -17,7 +17,7 @@ Prerequisites: JDK 17, update Intellij to the most recent version.
 1. Configure the project to use **JDK 17** (not other versions) as explained
    in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
    In the same dialog, set the **Project language level** field to the `SDK default` option.
-1. After that, locate the `src/main/java/Kipp.java` file, right-click it, and choose `Run Kipp.main()` (if the code
+1. After that, locate the `src/main/java/kipp.java` file, right-click it, and choose `Run kipp.main()` (if the code
    editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the
    below as the output:
    ```
