+++ docs/README.md
@@ -1,4 +1,4 @@
-# Kipp User Guide
+# kipp User Guide
 
 // Update the title above to match the actual product name
