+++ src/main/java/tasklist/EventTask.java
@@ -1,4 +1,4 @@
-package TaskList;
+package tasklist;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
